lots of refactoring, extended gdb stub, removed custom io, real software breakpoints instead of lookup, target features for gdb

+++ src/main/java/io/scriptor/arg/MemoryPayload.java
@@ -1,4 +1,4 @@
 package io.scriptor.arg;
 
-public record MemoryPayload(long size) implements Payload {
+public record MemoryPayload(int size) implements Payload {
 }
